Solution to Calculate and Print the Area of a Circle

+++ Calculate and Print the Area of a Circle@@6788d5012db25efec946f943/code.c
@@ -1,7 +1,7 @@
 #include<stdio.h>
 int main(){
-    int side;
-    scanf("%d",side);
+    float side;
+    scanf("%f",side);
     float area = 3.14*side*side;
     printf("Area: %.2f",area);
     return 0;
